+++ src/main/java/com/imsweb/naaccrxml/NaaccrXmlDictionaryUtils.java
@@ -956,7 +956,20 @@ public final class NaaccrXmlDictionaryUtils {
             _namespaceWritten = false;
 
             try {
-                writer.write("<?xml version=\"1.0\"?>" + System.lineSeparator() + System.lineSeparator());
+                writer.write("<?xml version=\"1.0\"?>");
+                writer.write(System.lineSeparator());
+                writer.write(System.lineSeparator());
+
+                if (BASE_DICTIONARY_URI_PATTERN.matcher(dictionary.getDictionaryUri()).matches()) {
+                    writer.write("<!-- This dictionary is maintained and provided by the NAACCR organization; it should not be modified.");
+                    writer.write(System.lineSeparator());
+                    writer.write("     If you need to define non-standard data items, please create your own user-defined dictionary.");
+                    writer.write(System.lineSeparator());
+                    writer.write("     Visit https://www.naaccr.org/ for more information about the NAACCR XML Data Exchange Standard. -->");
+                    writer.write(System.lineSeparator());
+                    writer.write(System.lineSeparator());
+                }
+
             }
             catch (IOException e) {
                 // ignore this one, the exception will happen again anyway...